+++ src/App.js
@@ -2,17 +2,17 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {Router, Route, IndexRoute, Link, browserHistory} from 'react-router';
 require('./index.css');
+import $ from 'jquery'; 
 import data from './data';
+import 'bootstrap/dist/js/bootstrap.js';
 import 'bootstrap/dist/css/bootstrap.css';
-import HomePage from './home';
-
+import 'bootstrap/dist/js/bootstrap.js';
 
 var App = React.createClass({
   render: function() {
     return (
       <div>
           <Nav />
-          <HomePage />
           {this.props.children}
       </div>
     )
@@ -23,23 +23,43 @@ var Nav = React.createClass({
   render: function() {
     return (
       <div>
-{ /*here
-    <nav className="navbar navbar-default">
-      <div className="container-fluid">
-        <div className="navbar-header">
-          <a className="navbar-brand" href="#">
-            <img alt="Brand" src="http://www.eisp.org.il/f-users/user_102867/website_103211/images/art-direction-png.png"/>
+
+     { /*FIRST NAV*/}
+    <nav className="navbar main">
+        <div className="logo">
+          <a href="#">
+           <img alt="Brand" src="http://adelantealliance.org/wp-content/uploads/2016/10/logo.jpeg"/>
           </a>
         </div>
-      </div>
-    </nav>
-*/}
+    
+        <div className="input-group">
+          <input type="text" className="form-control search" placeholder=" Quick Search"/>
+            <button className="btn btn-default go" type="button">Go!</button>
+        </div>
 
-{ /*here*/}
+
+        {/*ADDING SOCIAL MEDIA*/}
+         <div className="social">
+            <a className="media twitter" href="#">
+              <img src="http://icons.iconarchive.com/icons/limav/flat-gradient-social/256/Twitter-icon.png"/>
+            </a>
+
+            <a className="media facebook" href="#">
+              <img src="https://cdn0.iconfinder.com/data/icons/yooicons_set01_socialbookmarks/256/social_facebook_box_blue.png"/>
+            </a>
+
+            <a className="media instagram" href="#">
+              <img src="https://anneauweiler.files.wordpress.com/2014/04/icon-instagram.png"/>
+            </a>
+          </div>
+    </nav>
+
+{ /*SECOND NAV*/}
 <nav className="navbar navbar-default">
   <div className="container-fluid">
     { /* Brand and toggle get grouped for better mobile display*/}
     <div className="navbar-header">
+
 
        {/*TOGGLE BOTTON SHRINKS*/}
       <button type="button" className="navbar-toggle collapsed" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1" aria-expanded="false">
@@ -48,52 +68,41 @@ var Nav = React.createClass({
         <span className="icon-bar"></span>
         <span className="icon-bar"></span>
       </button>
-      <a className="navbar-brand" href="#">
-      Amazing Art
-       {/*<img id="logo" alt="Brand" src="http://www.eisp.org.il/f-users/user_102867/website_103211/images/art-direction-png.png"/> */}
-      </a>
+        <Link to="/">Home</Link>
     </div>
 
     { /*Collect the nav links, forms, and other content for toggling */}
-    <div className="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
+    <div className="collapse navbar-collapse">
       <ul className="nav navbar-nav">
-        <li className="active"><Link to="/">Home</Link><span className="sr-only">(current)</span></li>
-        <li><Link to="about">About</Link></li>
-        <li><Link to="/">New Art</Link></li>
-        <li><Link to="/">Collections</Link></li>
+        <li className="linetext"><Link to="/">Cat</Link><span className="sr-only">(current)</span></li>
+        <li className="linetext"><Link to="menu">Dragon</Link></li>
+        <li className="linetext"><Link to="about">Giraffe</Link></li>
+        <li className="linetext"><Link to="menu">Unicorn</Link></li>
+        <li className="linetext"><Link to="about">Walrus</Link></li>
 
 
         <li className="dropdown">
           <a href="#" className="dropdown-toggle" data-toggle="dropdown" role="button" aria-haspopup="true" aria-expanded="false">Search by category <span className="caret"></span></a>
           <ul className="dropdown-menu">
-            <li><a href="#">Action</a></li>
-            <li><a href="#">Another action</a></li>
-            <li><a href="#">Something else here</a></li>
-            <li role="separator" className="divider"></li>
-            <li><a href="#">Separated link</a></li>
-            <li role="separator" className="divider"></li>
-            <li><a href="#">One more separated link</a></li>
+            <li><Link to="about">Funny</Link></li>
+            <li><Link to="about">Eerie</Link></li>
+            <li><Link to="about">Weird</Link></li>
           </ul>
         </li>
       </ul>
-      <form className="navbar-form navbar-left">
-        <div className="form-group">
-          <input type="text" className="form-control" placeholder="Quick Search"/>
-        </div>
-        <button type="submit" className="glyphicon glyphicon-search"></button>
-      </form>
+      
 
 
        {/*RIGHT NAV*/}
       <ul className="nav navbar-nav navbar-right">
         <li><a href="#">Login</a></li>
-
+          
           {/*CART*/}
-        <li><Link to="menu" className="glyphicon glyphicon-shopping-cart">Cart</Link></li>
+        <li><Link to="menu">Cart</Link></li>
 
 
-
-
+  
+       
       </ul>
     </div>{ /*/.navbar-collapse */}
   </div>{ /*/.container-fluid */}
@@ -108,7 +117,7 @@ var Nav = React.createClass({
           <li><Link to="about">About</Link></li>
         </ul>
       */}
-
+        
       </div>
     )
   }
@@ -121,40 +130,27 @@ var Home = React.createClass({
   componentWillMount(){
     // this.setState({images:data.getImages()})
     // console.log('images', data.getImages())
-	},
+  },
   render: function() {
-
+    
     return (
       <div>
+      {/*BOOTSTRAP BOTTONS
         <h1>Welcome To Our Gallery Website</h1>
         <h4>BROWSE AROUND TO SEE WHAT WE HAVE IN STORE FOR YOU.</h4>
 
-
-         {/*BOOTSTRAP BOTTONS*/}
+         
         <button className="">Luis</button>
         <button>
           <span className='glyphicon glyphicon-thumbs-up'></span>
         </button>
-        <button type="button" className="btn btn-success">Success</button>
-
-         {/* BOOTSTRAP TABLE*/}
-        <div className="container">
-          <div className="row">
-            <div className="col-sm-4">
-              One of three columns
-            </div>
-            <div className="col-sm-4">
-              One of three columns
-            </div>
-            <div className="col-sm-4">
-              One of three columns
-            </div>
-          </div>
-        </div>
+        <button type="button" className="btn btn-success" className="tu">Success</button>
+        */} 
 
 
 
-          {/* BOOTSTRAP TABLE FOR ALL ITEMS*/}
+
+          {/* BOOTSTRAP TABLE FOR ALL ITEMS*/} 
         <div className="row">
           <div className="col-sm-6 col-md-4">
             <div className="thumbnail">
@@ -193,47 +189,115 @@ var Home = React.createClass({
   }
 })
 
+var MenuItem = React.createClass({
+    render: function() {
+    return (
+      <div className="m">
+        <div className="l">
+        <li><h4>{this.props.name}</h4></li><br/>
+          <li><img src={require(this.props.image)}/></li>
+          <p className="description">{this.props.description}</p>
+          <li>{this.props.price}</li>
+        </div>
+        
+      </div>
+    )
+  }
+})
+
+var Menu = React.createClass({
+  getInitialState(){
+    return ({menu:null,images: null})
+  },
+  componentDidMount(){
+    var gallData = data.getGallery();
+    this.setState({menu: gallData})
+  },
+    render() {
+    return (
+      <div className="gallery">
+
+       {/* <h1>Welcome To Our Gallery Website</h1>
+        <h4>BROWSE AROUND TO SEE WHAT WE HAVE IN STORE FOR YOU.</h4>
+
+
+         *BOTTONS
+        <button>
+          <span className='glyphicon glyphicon-thumbs-up'></span>
+        </button>
+        <button type="button" className="btn btn-success">Success</button>
+       */}
+          {/*SIDEBAR*/}  
+        <table className ="table-bordered">
+          <thead>
+            <tr>
+              <th>Deparments</th>
+            </tr>
+          </thead>
+          <tbody>
+            <tr>
+              <td>Category</td>
+            </tr>
+            <tr>
+              <td>Eerie</td>
+            </tr>
+            <tr>
+              <td>Funny</td> 
+            </tr>
+             <tr>
+              <td>Eerie</td>
+            </tr>
+            <tr>
+              <td>Funny</td> 
+            </tr>
+             <tr>
+              <td>Eerie</td>
+            </tr>
+            <tr>
+              <td>Funny</td> 
+            </tr>
+          </tbody>
+        </table>
+
+
+
+         {/*PRINTS ALL GALLERY*/}  
+          <div className="pictsConteiner">
+          <ul>
+           {this.state.menu ? this.state.menu.map((item, indx) => <li>{item.title}</li>) : null}
+           </ul>
+          </div>
+      </div>
+        
+
+      
+    )
+  }
+})
 
 var About = React.createClass({
   getInitialState(){
     return({about:null})
   },
   componentWillMount(){
-    this.setState({about:data.description})
+    this.setState({about:data.getAbout()})
   },
   render: function() {
     return (
       <div >
         <h1 className="modal-header">About Luis</h1>
-        <p className="vertical-container"></p>
-        <p></p>
-        <p></p>
+        <p className="vertical-container">{this.state.about.restaurant_name}</p>
+        <p>{this.state.about.restaurant_slogan}</p>
+        <p>{this.state.about.restaurant_slogan1}</p>  
         <h2 className="vertical-container">Contact</h2>
-        <p className="main"></p>
-        <p></p>
-        <p></p>
+        <p className="main">{this.state.about.restaurant_address}</p>
+        <p>{this.state.about.restaurant_email}</p> 
+        <p>{this.state.about.restaurant_phone}</p>
       </div>
     )
   }
 })
 
-var MenuItem = React.createClass({
-  getInitialState(){
-    return ({product: null})
-  },
-    render: function() {
-    return (
-      <div>
-        <ul>
-        <li>{this.props.name}</li><br/>
-          <li><img src={require(this.props.image)}/></li>
-          <li>{this.props.description}</li>
-          <li>{this.props.price}</li>
-	    	</ul>
-    	</div>
-    )
-  }
-})
 
 
 
@@ -253,6 +317,7 @@ ReactDOM.render(
     <Route path="/" component={App}>
       <IndexRoute component={Home} />
       <Route path="about" component={About} />
+      <Route path="menu" component={Menu} />
     </Route>
     <Route path="*" component={NotFound} />
   </Router>,
